Add creation_facility to Appointment
To track patient transfers, we will continue using facility attribute
on the Appointment model and add a creation_facility attribute that will
track the original facility that it was transferred from.

The overdue list and the SMS reminders will continue to use the facility
attribute to perform and no other functional change needs to be made.

i.e. This is 100% backwards-compatible