Add My Facilities Dashboard  (#710)
This adds the new "My Facilities" overview page. The "My Facilities" dashboards are still a work in progress, but the overview page is releasable as-is. UI navigation links to the new dashboards will remain hidden for the time being, controlled by a feature flag which can be flipped when more dashboards are available.

Co-authored-by: Prabhanshu Gupta <guptaprabhanshu1@gmail.com>
Co-authored-by: Daniel Burka <danielburka@gmail.com>
Co-authored-by: Akshay Gupta <kitallis@gmail.com>