Consolidate high risk (#759)
Story: https://www.pivotaltracker.com/story/show/170594701

## Overview

This PR updates risk levels of BP patients. In order to do so, it also updates the levels of hypertension that a BP reading can be categorized in. Read on to find out what definitions are proposed in this PR. The reference rules for these changes is shown below.

---
![image](https://user-images.githubusercontent.com/4241399/72529937-4f998d80-3894-11ea-81bb-f9f8f7b6c0f1.png)