Hide non-IN support items (#796)
Story: https://www.pivotaltracker.com/story/show/171181270

This PR hides the support section from the help screen for BD (ie, any non-India) country by tagging relevant elements with a country class and hiding through CSS.

This approach is a lot cleaner than the attempt made in #795 . I think the long-term approach is to make these translation files ERB-like templates so we can introduce the necessary logic to hide/show things directly into the template itself. Today, the impact of such a solution on Transifex is unclear, so we're going with a surgical CSS approach instead.