My facilities permissioning (#848)
**Story card:** https://www.pivotaltracker.com/story/show/171752073

## Because

We want to grant CVHOs access to My Facilities one by one.

## This addresses

Removes the use of the My Facilities feature flag, and instead relies solely on the provisioning of the `view_my_facilities` permission. No users will be pre-populated with this permission.

The permission will also _not_ be added to any of the role presets to ensure that all existing users' roles are preserved in the Permissioning UI. Eg. Super admins will remain super admins, and will remain able to invite new super admins to Simple.

This PR also addresses a couple of bugs found along the way:
* Inactive facilities were not scoped correctly on the overview page - a523025
* Facilities without zones were not scoped correctly on the BP Control page - a665565