Add app.json (#870)
**Story card:** https://www.pivotaltracker.com/story/show/172145441

## Because

We want to set up Heroku Review Apps

## This addresses

The first step to review app configuration is the addition of an `app.json` file that governs how to set up a new Heroku app. This is very similar to, say, a terraform config. It dictates environment variables, Heroku buildpacks, required addons, processes.

[Heroku reference](https://devcenter.heroku.com/articles/app-json-schema)