Small formatting fix in HELP and added TOS/Privacy links. (#868)
Fixed the side scrolling in HELP I think. Also, added the TOS and Privacy to Help. They will later end up in Settings, but it's important that they are in the app asap.

**Story card:** NONE

## Because

The Help is inheriting some padding from the common_webview css and I had to disable it.

## This addresses

This also includes 2 links to the website of Simple for the TOS and Privacy. Will these open in-place or in Chrome? Either is probably fine.