Upgrade our Follow-ups queries across dashboards (#888)
We've seen issues of follow-ups not lining up between the old dashboards, within the old dashboards and across Progress Tabs. This is primarily due to different ways of calculating follow-ups and not accounting for all edge-cases properly.

This attempts to simplify the ergonomics of calculating follow-ups and the queries themselves. Since most user-facing dashboards are cached, the queries are intentionally not optimized to be very fast. However, they are reasonably fast for most cases (< 2-3s, worst-case).