Add indexes to improve patient_summaries performance
These two indexes cut down significantly on time spent in two `order
by` clauses we are doing for the patient_summaries view.

These indexes are compound and need to sort the second date field descending
to match how we commonly retrieve appointments and blood_pressure.

On the main overdue list page this cut the overall page load time in
half.