Freeze time to 10 am to avoid end of month issues
I'm not entirely sure what is going on here, but it appears since we are
on the last day of the month these specs started failing that are
setting up fixture data on one month intervals.

In the interest of getting CI green, it looks like freezing time to the
morning here fixes things, at least locally.  So lets see if that helps
CI.