Spike out adding risk level to patients
Adds a small service class and a data migration to add risk level to
Patients. The next step will be a daily Sidekiq job to update this risk
level for patients.

I ran this locally with sandbox data, and this totally naive strategy
took about 10 minutes to update all patients.