Upgrade to rails 5.2 (#915)
* Alphabetize Gemfile

Makes things easier to read, track and diff long-term

* Start upgrading to Rails 5.2

Had to do

```
bundle update rails anonymization
```

to get bundler to play nice

* Add bootsnap; merge one other small change from 5.2

via rails app:update

* Add new framework defaults for 5.2

* Dont change cookie strategy for now

* Use the older bundler

* Use dynamic attributes for factory-bot

This addresses a deprecation in factory bot that full on breaks in Rails
5.2.
You can see the details on the deprecation here:

https://thoughtbot.com/blog/deprecating-static-attributes-in-factory_bot-4-11

* Moved SCSS partials to a subdirectory to make sprockets 4 happy

See issues arising here:
https://github.com/rails/sprockets/blob/master/UPGRADING.md#manifestjs

* Stay on factory bot 4.8.x

The upgrade to factory bot 5 is breaking specs in this branch, making it
hard to figure out what is rails related breakage and what is Factory
Bot related. Lets stay on Factory Bot 4.8 for now to better isolate our
upgrades.

* Lock to factory 4, don't go to 5 yet

* Fix patient_business_identifier spec

Having two validations of the same field and type probably doesn't do
what we want it too, especially in Rails 5.2 and newer. I don't know if
they have any sort of 'merge' behavior.

I unrolled to a custom validator, and also unrolled the specs a bit for
clarity.

* Update user agent to match 'http' gem version

* Update more user agents to match 'http' gem version

Co-authored-by: Tim Cheadle <tcheadle@resolvetosavelives.org>