Use form attribute so pagination / facility filters shares state:
This overrides the facility / pagination select forms so that they post
to the search filters form when on the Overdue listing. This means
everything goes through the same form for this page, which makes
managing the state much easier, and we don't need hidden fields leaking
into both forms.

This also fixes up the filters to work more better in the mobile view.