Search filters for overdue patients (#907)
* Add Guard for quick feedback during dev

This adds https://github.com/guard/guard-rspec, which wire up guard and
guard-rspec to get continual, fast testing while working locally.

* Add Guardfile

This is all generated by the guard-rspec setup, we can modify it as we
see fit to match our project and conventions.

* Try to get a failing test going for overdue filter

* Wire up filter for showing all overdue

* Turn on focused specs

Allows running just specific specs by changing "it" to "fit" in rspec

* Add new layout for overdue listing

Need this for the new sidebar with search filters

* Wiring up more of the filters

* Dont need this filters

* Updates to bring in line w/ mockup

Removing unused filters; adjusting copy

* Trying to preserve state between patient and facility filters

This is tricky because they are in different forms, so we need to use
hidden fields to keep the state.

We should probably just merge the forms if at all possible here.

* Get state working across facility filter and search filters

Ugh this was gnarly.  Forms that are GETs ignore the query string in
their action, so we *have* to pass around the state via hidden field
tags.

* Extract search logic

This all should be another object, but not until we know where else
these filters are going to be used

* Ensure we have an Array

* Add back the hidden submit

* Ensure we have an array here

* Default to only showing less than one year overdue

* Only allow downloads when facility is selected

* Set window size to a desktop size:

This is to ensure that our dashboard feature specs operate in 'desktop'
mode, which _I think_ is the version we are testing more commonly.

We may want to introduce a way to switch between this and a mobile
chrome if we have tests that exercise both layouts.

* Handle the hackiness around a default checkbox and search filters

* Fix risk level query

* Add indexes to improve patient_summaries performance

These two indexes cut down significantly on time spent in two `order
by` clauses we are doing for the patient_summaries view.

These indexes are compound and need to sort the second date field descending
to match how we commonly retrieve appointments and blood_pressure.

On the main overdue list page this cut the overall page load time in
half.

* Extract PatientSummaryQuery

* Fix the spec to really assert things; linting

* Remove dead code

* Add bootsnap to speed up startup time

This shaves a second off an individual test run, which is pretty nice
locally.

This is already in the Rails 5.2 upgrade but I'd like to ship it
independently because its a big win locally.

* Maintain state with pagination/facility submission as well

And simplify the default search filter logic in appointments controller

* Remove unnecessary comment

* Spec fixes from code review

* Fix var name in spec

* Fix boot.rb conflicts

* Use form attribute so pagination / facility filters shares state:

This overrides the facility / pagination select forms so that they post
to the search filters form when on the Overdue listing. This means
everything goes through the same form for this page, which makes
managing the state much easier, and we don't need hidden fields leaking
into both forms.

This also fixes up the filters to work more better in the mobile view.

* Use the human readable label

* Extract filters to object w/ most knowledge of them

* Handle case for both phone number filters

these cancel each other out, effectively

* Fix spec so it sets submitted value

* Reuse all_overdue scope for overdue:

These should really be renamed soon if we agree on the "lost to follow
up" terminology