Prevent more deprecations (hopefully) (#980)
* Dont mark follow_ups_with as private

This causes all sorts of deprecation warnings, and I don't see a way
around this without relaxing the method visiblity here:

The warnings are :
```
DEPRECATION WARNING: Delegating missing follow_ups_with method to Patient. Accessibility of private/protected class methods in :scope is deprecated and will be removed in Rails 6.0. (called from block in <class:Patient> at /Users/rsanheim/src/simpledotorg/simple-server/app/models/patient.rb:67)
```

* Raise on deprecations in dev and test

If we are able to get a green build w/ this set, we should turn it on
for dev and test to keep us honest.