Add a whitelist for logging:
This ensures we only log keys that we explicitly want, and everything
else is filtered. This lets us log just what we want for tracing /
debugging (so things like UUIDs, request times, HTTP info) and all
personal information will be dropped.

Also adding a request spec that is basically the patient sync spec that
verifies that patient attributes were _not_ logged after the fact.
Getting a working spec on the logging was tricky, and it still needs
some work to properly verify things.