Dynamic manifests (#986)
**Story card:** https://www.pivotaltracker.com/story/show/172912962

## Because

We need to teach review apps to serve meaningful manifests, because #985 will start creating working app builds to test against review apps.

## This addresses

Serving dynamic manifests in the `development` and `review` environments. The only dynamic bit is the API endpoint listed against each country in the manifest.

The inclusion of the `development` env in this PR is to facilitate local server + emulator setup. The relevant details have been added to the README as well.