Lockout a user after five failed attempts at logging in (#975)
* Extract PhoneNumberAuthentication::Authenticate

This is to prepare for enhancing security around phone number based
authentication. The controller is doing a bit too much here, so
extracting to an Authenticate object that handles the actual
verification and access token type things.

* More refactoring; simplify specs in controller ...

now that we are covering much of that functionality in the authenticate
model

* Make sure we invalidate the OTP after login

* Change multi-branch if/else to case statement

* Further cleanup and simplification

* newline & whitespace

* Add failed_attempts and locked_at to phone_number_authentications

* Schema updates for api lockouts

* Add spring; cuts about ~3 seconds off every local test run

* Turn off the annoying sidekiq message in tests

* Use binstub for guard

* track failed attempts and set locked_at when we hit 5

* Tell the user how long they are locked out until

* Handle the unlock path after 20 minutes

* Fix failing spec:

merge in a password to params, as this spec is only testing the email
validation

* Extract constants; rename a local for clarity

* Reset failed attempts on successful login

* Set a specific time to avoid flakey tests:

If we don't do this, we can get failures due to Ruby's times holding
more precision than Postgres times. Ruby stores time tracked to the
nanosecond, while Postgres just goes to microsecond.

If we set an exact time via Timecop.freeze, we don't have to worry about
fractional seconds.

* Set to a specific time to avoid flakey tests

* Validate failed_attempts; move unlock method

* Move lockout settings to env vars

* Move some logic to PhoneNumberAuthentication

* Use predicate matcher

* More concise language per @dburka

* Update tests for new messaging

* Move track behavior to model

* Set auth lockout vars to defaults with optional ENV override

* Add tests for phone_number_authentication lockout logic

* Better specs

* Apply standardrb formatting