Add flipper memoization; add all permissions to `owner` (#1052)
* Configure Flipper Memoizer

* Move middleware to flipper initializer

* Validate we have a real permission slug

* Add a sanity test for our Permission hashes

We have a convention that the name of the permission matches the slug,
so we should enforce that to prevent any drift over time

* Add view_flipper_ui and view_sidekiq_ui to owner access level:

If we don't do this, new admins created in dev won't have all the
correct access levels, leading ot a very confusing state. There doesn't
seem to be a way to add new access levels to an admin, you can only
remove them. So for new admins we should really be adding all the
possible access levels.

* Move permissions spec to correct dir