Moar help screen translations (#1057)
**Story card:** https://www.pivotaltracker.com/story/show/173350287

The next (and perhaps last) batch of help screen translations. Adds translations for new help screen designs to:

* Hindi
* Tamil
* Telugu

As with #1027, I have left the unused files in the `upcoming` directory untouched for now. Once we merge this PR, I'll open a subsequent PR to clean up the `upcoming` directory, and the corresponding resources from Transifex.