Locale consolidation (#1097)
**Story card:** https://www.pivotaltracker.com/story/show/173528192

## Because

Our locales are split up into files more than they need to be.

## This addresses

Consolidation of locales as appropriate. The `api` locales are left untouched because, they are strings exposed to the client, that are translated in all the Simple App languages. To preserve our translation workflow, they are untouched.

The commit trail lists the changes made fairly clearly.

More notes:
* Reduced Sidekiq concurrency in review apps drastically
* Fixed an incorrectly escaped newline in the OTP SMS across languages
* Added a missing quotation mark in "genders" across languages.

I diffed the full locale hashes between `master` and this branch to confirm that no other changes were introduced.