Retrieve all registration counts with one query:
This works well in my test and gets all the counts, but fails in the
browser due to some time zone issues going on.

I tried to use the date_to_period_sql but was struggling, as that still
didn't seem to solve things.

I may need to dig into
https://phili.pe/posts/timestamps-and-time-zones-in-postgresql/ to
understand how timezones are working with PostgreSQL here.