Allow admins to set any permission for useres in admin UI
Not having this set makes it impossible to turn on new permissions as
they get added, which is very annoying and going to be a blocker for
designers or others who need to turn on Flipper access locally.

I dunno the consequences of this in terms of security, but this seems
like a good change to me?