Revert the changes that caused the ordering bugs (#1109)
**Story card:** https://www.pivotaltracker.com/story/show/173508399

This PR addresses a bug introduced in #1085 , where some necessary explicit ordering was removed, attempting to rely on ordered associations. However, due to the batching that occurs in the patient exporter, when eager loading ordered associations, the ordering is not respected. Hence we abandon eager-loading ordered associations, and instead make explicit ordered queries iteratively instead.