Use groupdate to get totals, and use Ruby to sum the running total
This gets rid of a pretty complicated SQL query where we also have to
worry about handling time zone conversion ourselves.

The ruby here is pretty straightforward and will be easier to maintain
long term.

This also changes the interface to take the district instead of a list
of facilities, which is really the object we care about.