Use RequestStore so we can easily bust caches from controller layer down
Without this we had to pass state all around the different service
objects at play, which was annoying and not really worth the hassle.
RequestStore provides store local to the request that is reset
thread-safely after every request, so that is a decent place to store
this very small piece of state.