Spike: Simple User Roles
Adds a single model - Role - which belongs to a User and a Resource. The
Role defines whether a User is a super admin, admin, or analyst of a
resource. Right now a resource can be an Org, FacilityGroup, or
Facility.

I think this is the simplest possible data model that could work, and I wanted to give
it a shot to see how straightforward we can make our authorization.

I also added authorization checks via a Flipper feature flag
for a couple controllers to show how nice using this is.
With this approach we are using the database to handle as much of our
authorization as possible or us. This is done via a couple UNION queries to capture the
hiearchial nature of the roles, so we can get all facility groups
or all facilities a User may have access to as an ActiveRecord relation.

That means we can then chain off of that relation so that our auth code
can stay dumb and simple. Controllers should scope off `current_admin` to ensure we
are only finding resources a user should have access to.
Using this approach means resources that a user doesn't have access to
should never even be loaded from the database for a properly written
controller.

If a logged in admin tries to load some facility they shouldn't have access to, our controllers
will raise RecordNotFound errors, because we can now do things like
this:

```
current_admin.facilities.find_by!(slug: params[:slug])
```

Or read only access:

```
current_admin.facilities(role: :read).find_by!(slug: params[:slug])
```

Or finding a facility from a facility group, for nested controller
cases:

```
current_admin.facility_groups.find_by!(slug: params[:slug])
```

Or finding all facilities a user has access to:

```
current_admin.facilities
```

I realize this may not handle some more complicated edge cases, but wanted
to get it out there for feedback and thoughts before considering all the
trickier scenarios. Feedback welcome!