Bugfix: CSV formatting (#1176)
There appears to be problems escaping characters in the overdue list CSV which causes column data to bleed into other columns. Cohort downloads also has a quote escaping issue in the column headers.

This fixes the CSV generation in these downloads. Since these CSVs are rendered via ERB, templating a row sanitises special characters. For example " becomes a &quot; which some CSV clients handle correctly, some can't. Strings with a comma in between are read as separate cells because they are not wrapped in ".

This sanitisation is desirable in HTML views since they are executed in the browser. File downloads should contain the raw strings. Running the string through an html_safe spits the correct unsanitised strings into the CSV.