Clean up some javascript errors:
All of this JS is firing on every page...I'm not sure if this happened
at some point accidentally w/ a rails upgrade due to how Sprockets
works?  In any case, we get errors on every page because we are loading
JS that isn't valid for them.

I'm adding some element checks here to remove the errors, but there may
be a better way to do this. Paging @kitallis and @harimohanraj80,
because I know they both worked on some of the user analytics stuff
which is for the Progress tab.