Do not memoize existing_patient and avoid race condition (#1177)
We ran into this bug with the recent patient merge refactor. This happens due to a race condition when two syncs simultaneously try to create the same patient.

In MergePatientService, caching the existing_patient results in a case where one sync creates a patient and the other sync fails:

- Both requests fetch and memoize existing patient data (which is nil)
- Request 1 (faster) computes merge status, determines it's a new resource and creates it
- Request 2 (slower) computes merge status and determines it's an update. It tries to use the memoized existing_patient for fetching the metadata but fails.

This removes the memoization on existing_patient to avoid this race condition.