Refactoring: preparing the way for easier quarterly cohorts and monthly cohorts (#1183)
* Start wiring up quarterly reports

Mostly brute forcing this before seeing how we can refactor things

* Add view_component https://github.com/github/view_component

* Add a Period object that can holds a Quarter or Time/Date (for months)

Trying to get a handle on all the period selection our reports have to
do.

* Clean up some javascript errors:

All of this JS is firing on every page...I'm not sure if this happened
at some point accidentally w/ a rails upgrade due to how Sprockets
works?  In any case, we get errors on every page because we are loading
JS that isn't valid for them.

I'm adding some element checks here to remove the errors, but there may
be a better way to do this. Paging @kitallis and @harimohanraj80,
because I know they both worked on some of the user analytics stuff
which is for the Progress tab.

* Moving to use Period object for reports

This simplifies the disjoint nature between months (represented by
dates) and quarters. Still WIP but getting better here.

* Handle point in time period calcs for benchmarks

Need to handle a single Period vs a range of Periods

This lets us handle the benchmarks that TopRegionService calculates

* Fixing specs

* Remove data, using Period; fixing tests and cleaning up

paving the way for proper quarter cohorts

* Grab correct registrations count for quarterly cohorts:

keeps around the period registrations and the cumulative registrations

* Getting the cumulative regs vs period registrations working

* cleanup

* Simplify

* bump cache; fix quarter check

* Send back cumulative_registrations as the registrations

* linting

* Fix tests

* extract utility method

* Use question method

* Move cache logic to keep consistent

* alpha order