Use group_by_period to get ALL registration counts:
This ensures we are using a consistent method to get registration
counts. Previously we were using grouped counts only for the ranges the
client is calling for, which was making things tricky because we needed
to also know previous registrations to get the cumulative counts right.

So now we get counts using the same query, and then we trim the result
sets we return to match the periods the client is asking for.

This should help ensure we get the correct registration counts, which is
critical as these numbers are the foundation for determining all the
other dashboard figures.