Do not filter by size if all size filters are selected (#1270)
**Story card:** https://app.clubhouse.io/simpledotorg/story/979/my-facilities-registrations-don-t-include-facilities-without-a-size

Facilities without sizes set are excluded from My Facilities pages because of the default size filtering that takes place. This is a quickfix to address the problem.

**If all sizes are selected, don't filter by size.**

This ensures that facilities without a size are also included in these lists.

We could be stricter with data validation, and fill in missing sizes in production, etc, etc, but I'd rather let administrators take care of that aspect of data hygiene themselves.