Make the from_time default to something deterministic
We need this to be something more stable to use it as our cache key,
otherwise, every cache fetch would be a miss due to time.current.

Also, I don't think having the from_time be the current time is actually
meaningful, as its only used as the basis for determining the
offset_date, which is a 'months' value anyways.