Handle nil access_levels:
We need to handle this for the transition period before everyone has an
access level.  I turned this up by turning on permissions_v2 for myself
in dev, and hit errors when trying to manage admins in my local
instance.

Also changing UserAccess to a proper object, because its easier and
has a ton of behavior on it anyways.  note that we were actually
_subclassing_ Struct previously, which is generally recommended against
...but is an easy mistake to make w/ structs.