Refactor feature check responsibility to user for actor flipper checks (#1309)
* Push feature check responsibility to user

This makes for more readable code throughout all our feature checks, and
also guides callers to use the one way of checking actor gates on our
current_admin / current_user.

* Handle nil access_levels:

We need to handle this for the transition period before everyone has an
access level.  I turned this up by turning on permissions_v2 for myself
in dev, and hit errors when trying to manage admins in my local
instance.

Also changing UserAccess to a proper object, because its easier and
has a ton of behavior on it anyways.  note that we were actually
_subclassing_ Struct previously, which is generally recommended against
...but is an easy mistake to make w/ structs.

* Use _v2 for the new methods as well

more consistent, and emoji in method names is a bit too clever by half
😏

* rename new authorize method to _v2 for consistency

easier to grep for when everything follows the same naming convention