Migration script to move from old user permissions to new (#1312)
Add a script to create both User#access_level and Access using old permissions from user_permissions.

This script runs for a default organization name of IHCI but it can be overridden and should work for Bangladesh as well. This script does not migrate admins who don't fall into a preset category of permissions (custom permissions).