Enable Rails 5.2 Content Security Policy (#1417)
Our app doesn't expose a Content Security Policy (CSP).

Create one for all pages (for the time being) and configure any third-party hosts as necessary.

Sentry has a neat feature that can be used to report CSP violations, we have set our CSP to report-only mode and which should show up as alerts on Slack if anything is violated. This will allow us to find any violations that we may not catch immediately and can fix them incrementally, rather than outright breaking the loading of the page.