Remove fakeredis
Fakeredis inserts itself in the way in a very obtrusive way as the
'default driver' once it is loaded, which makes it hard to deal with in
terms of testing. See
https://github.com/guilleiguaran/fakeredis/issues/63 for more details.

There are alternative gems that we can more easily load only where
needed that aren't so invasive.

This is also step 1 towards removing our redis gem for caching, because
we can use Rails builtin RedisCacheStore as of Rails 5.2