Simplify the syncing codepaths (#1452)
There's a bunch of logical redundancy and cruft around how the sync mechanism works; makes it hard to understand the codepaths.

Remove the PrioritisableByFacility mixin in favor of simply delegating it to SyncToUser and make way for easily modifying the sync_region for all the sync resources in a generic way.

Additionally, we've modified how current_facility_records and other_facility_records are synced. They are synced scoped by the patients in the current_facility, rather than the facility itself. This is conceptually more accurate (albeit equivalent in terms of data to the previous change we made in #1375).

Co-authored-by: Prabhanshu Gupta <guptaprabhanshu1@gmail.com>