Ensure we only ever init a path in very explicit case
Overriding parent on an ltree backed model feels dangerous and
surprising, and the 'reparent' case should be a very clear exception and
not the rule.

We can use a virtual attribute to setup a clear path where a Region
should change its path, and for all other cases we can assume the path
should not change. This is a more stable way to go forward w/ regions
because we will be able to find easily where the path should change:
whenenver we do set `reparent_to: object` for a region. The only cases I
can think of this are:

1) when a Region's slug is changing (which shouldn't happen right now
anywhere, as we have no code paths that change a slug)
2) when a Region is first being created, either via the backfill or on a
Regions first creation via a callback