Upsert Regions through source models – I (#1486)
* Add some basic region callbacks for the all RegionSource models

* Fix some issues in the RegionBackfill script

* Thanks standardrb

* Revert some additional, but unnecessary standard changes

* Write a better Organisation region callback

* Keep the parent API (phew)

* Fix spec

* Cleanup parent methods

* Remove Region callbacks from FacilityGroup and Facility

* Remove update_children callback from Region

- This is not necessary as it will be superseded by using slugs for path

* Remove the spec helpers

* Fix specs

* Use enums to pull the right region type

* Create root only if root exists

* Reorder Region

* Fix root presence check

* Add data migration to create Root region

* Fix specs

* Create root region in seed

* Add back spec helper

* Fix organization spec

* Don't copy over source slug and avoid weird edge cases

* Fix slug specs

* Typo

* Limit the scope of when org / region callbacks fire

we only need to do an update if the name or desc has changed

* Ensure we only ever init a path in very explicit case

Overriding parent on an ltree backed model feels dangerous and
surprising, and the 'reparent' case should be a very clear exception and
not the rule.

We can use a virtual attribute to setup a clear path where a Region
should change its path, and for all other cases we can assume the path
should not change. This is a more stable way to go forward w/ regions
because we will be able to find easily where the path should change:
whenenver we do set `reparent_to: object` for a region. The only cases I
can think of this are:

1) when a Region's slug is changing (which shouldn't happen right now
anywhere, as we have no code paths that change a slug)
2) when a Region is first being created, either via the backfill or on a
Regions first creation via a callback

* Clean up linting issues

* Add some comments for clarity; clear out the reparent_to state

* maintain the same slug for FGs / facilities when we can

this will make things easier for region reports so that they can
maintain the same URLs after transitioning to Region based reports

* Dont try to log validitity in DryRunRegion

this causes weird side effects

Co-authored-by: prabhanshuguptagit <guptaprabhanshu1@gmail.com>
Co-authored-by: Rob Sanheim <rsanheim@gmail.com>