Dont override slug generation for factory created facilities (#1501)
* Dont override slug generation for factory created facilities

This overrides what FriendlyId would do by default, which is not
something we want to mess with in our test data. It also obscures the
default behavior of FriendlyId where it appends a UUID for unique slugs.

* Dont override slug generation for factory created facility groups

This overrides what FriendlyId would do by default, which is not
something we want to mess with in our test data. It also obscures the
default behavior of FriendlyId where it appends a UUID for unique slugs.