Region backfill - better handle slug dupes (#1502)
* Dont override slug generation for factory created facilities

This overrides what FriendlyId would do by default, which is not
something we want to mess with in our test data. It also obscures the
default behavior of FriendlyId where it appends a UUID for unique slugs.

* Dont override slug generation for factory created facility groups

This overrides what FriendlyId would do by default, which is not
something we want to mess with in our test data. It also obscures the
default behavior of FriendlyId where it appends a UUID for unique slugs.

* Handle blocks and facilities with same slug for RegionBackfill

We can continue to try to keep slugs consistent while ensuring they
remain unique if we use the built in friendly_id set_slug method.

This method will take into account our slug as a candidate but fallback
to the unique slug generation as necessary.

* Define Region.root to return the _one_ root level region

we should never have multiple root regions, so we can simplify this by
returning the one record

* Clean up: prefer block name; dont need .first for Region.root

* No need for .first, just grab first region