Remove unassociated facilities (#1524)
We have a bunch of unassociated facilities that are unused. They were originally associated but disassociated later because of bad name or something else. This feature is not useful anymore since previously facilities used to be created before FGs, this is not the case anymore and now this data will only adds more entropy to our move towards Regions.

Remove unassociated facilities from India:production and remove the feature on the view layer as well, just show the list of associated facilities.