Rename region methods to be more explicit
Because we sitll have source models like Facility and FacilityGroup, we
should name the methods that can move up or down the region tree very
explicitly ...other wise its very confusing if you see a call like this:

  region.facilities

Does that return actual Facilities, or Regions of type facility?

This changes the ancestor methods to have a _region suffix and the
descendant methods to a have a _regions suffix. It also gives the scopes
provided by the enum a _regions suffix for consistency.