Overhaul seed data generation (#1503)
* Dont override slug generation for factory created facilities

This overrides what FriendlyId would do by default, which is not
something we want to mess with in our test data. It also obscures the
default behavior of FriendlyId where it appends a UUID for unique slugs.

* Dont override slug generation for factory created facility groups

This overrides what FriendlyId would do by default, which is not
something we want to mess with in our test data. It also obscures the
default behavior of FriendlyId where it appends a UUID for unique slugs.

* Handle blocks and facilities with same slug for RegionBackfill

We can continue to try to keep slugs consistent while ensuring they
remain unique if we use the built in friendly_id set_slug method.

This method will take into account our slug as a candidate but fallback
to the unique slug generation as necessary.

* Define Region.root to return the _one_ root level region

we should never have multiple root regions, so we can simplify this by
returning the one record

* Clean up: prefer block name; dont need .first for Region.root

* No need for .first, just grab first region

* Simpler, faster seed task

Just generating patients & blood pressures right now

* Create more facilities and districts; simplify this seed code a bit

* Bit less patients range

* Get more variable facility data:

set a minimum registration data at the facility level, so that we get
different windows of data per facility.

Also add some more db tasks for making all this stuff easier

* moar data

* Notes from pairing with @dburka

* Scale patients to create based on facility size

* remove dead config

* Better weighted genders for seed data

* create bps, encounters, and observations via AR import

* Make sure we find a gender

* Create a possible larger set of facility groups

also use save! in user factory

* Ability to adjust scale factor via ENV var

* linting

* Move seed runner to its own namespace under lib

* linting

* Different performance ranks for facilities; add appointments

* refactor and more tests

* Sum the totals across all facilities

* Create config object loaded from env vars; use that to simplify

This allows us to have a fast data set used for test / dev, and a large
dataset to use elsewhere.  Also allows us to toggle random counts
on/off, since we don't really want that turned on for tests.

* Split out BP seeder; bit more testing

* fix test

* Create Facilities and Users via AR import

* Create users and phone number auths via AR import

* Less BPs per user; more benchmarking and real data

* Use truncate to clear patient data much faster; clear mat views also

* Deterministic appointments in test mode

* better output

* fast seed doesn't need a scale factor; less users

* Add user seeder with spec:

Also ensure we get some correct counts for the whole run

* less facilities for now

* Less BPs per patient, 50 is unreasonable

* Run our seeds for rails db:seed; linting

* skip Facility slug duplicates

handling this will require some fancier callback firing for friendly id,
which is really not worth it right now

* Rename for clarity; count created patients from the result

* Weighted sample of facility sizes for more accurate distribution

* Add some basic progress indicator

* Do the patient seeding in parallel

* Don't do things in parallel in test

* less facilities

* Clean up; ensure records are valid

* rename

* fix spec

* split out seed types to make it easier for different dataset sizes

* Process all facilities in on parallel batch

* clean up

* remove constants

* fix test state

* Dont need this scale factor

* Remove dead seed code

* remove dead code

* need these constants

* More facilities in large; fix facility seeder names

* Better progress; make sure mat views get refreshed after seed

* Track elasped time; refactoring

* more clean / refactoring

* Keep results isolated to their facility:

to avoid any issues or confusion with the parallel processing

* Add fake names; start using that for village

* Use a standard org for dev seeds; Stop using IHCI as name

* Use the fake district names

* fix specs

* fix names

* explain this

* linting

* Need to configure for review apps as well

* Ensure hypertensive factory generated patients dont have 110 diastolic

The problem w/ 110 diastolic is that some criteria in our app can make
that flip in to a higher risk level (like the PatientSummary
view)...which was causing specs to intermittently fail.

Setting upper bound to 109 will fix that.

* Seed data overhaul - testing Heroku review apps (#1545)

* Try using json logger for Heroku
* Check env for when this runs on Heroku
* Dont use Datadog on Heroku
* fix typo
* push commit to deploy review app
* Need gems for seed in all environments

  This is necessary for sandbox / heroku environments
  Also remove the old task from the heroku setup

* Need to require progressbar
* Remove commented code
* Remove debugging