Refactor Facility CSV upload  (#1554)
Became very unwieldy to add simple new validations on Facility (or CSV upload).

It pulls out the CSV parser and CSV-specific validations into their own classes.

The validations need to validate both on a collection and on a per-row basis, they are both handled within the same class. The model validations are also triggered from within for good measure.